requirements.txt
customtkinter==5.2.2
websocket-client==1.7.0
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.1
pyotp==2.9.0
schedule==1.2.1
pyinstaller==6.3.0
